avoid sending user errors to Sentry
