+++ src/error/abstract-error.js
@@ -1,28 +1,10 @@
 /** @flow */
-import R from 'ramda';
-
-export const systemFields = ['stack', 'code', 'errno', 'syscall'];
 
 export default class AbstractError extends Error {
+  userError: boolean; // user errors are not reported to Sentry
   constructor() {
     super();
     this.name = this.constructor.name;
-  }
-
-  // partially forked from 'utils-copy-error' package
-  clone() {
-    const error = this;
-    // $FlowFixMe
-    const err = new error.constructor(error.message);
-
-    systemFields.forEach((field) => {
-      // $FlowFixMe
-      if (error[field]) err[field] = error[field];
-    });
-    Object.keys(error).forEach((key) => {
-      // $FlowFixMe
-      err[key] = R.clone(error[key]);
-    });
-    return err;
+    this.userError = true;
   }
 }
